Refactored Nav class
