+++ src/main/java/gov/nci/WebAnalytics/MegaMenu.java
@@ -11,7 +11,6 @@ import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import com.nci.Utilities.Nav;
 import gov.nci.framework.PageObjectBase;
 
 public class MegaMenu extends PageObjectBase{
