+++ src/main/java/gov/nci/WebAnalytics/Nav.java
@@ -1,8 +1,8 @@
-package com.nci.Utilities;
+package gov.nci.WebAnalytics;
 
 import gov.nci.Utilities.ConfigReader;
 
-public class WANav {
+public class Nav {
 
 	// New ConfigReader
 	private static ConfigReader config = new ConfigReader();
